Relay Lists: - consolidate common code - statically sort list when editing to prevent rearrangement due to editing - revert to user sorting after editing - scroll to center entries before and after editing
